Improved the compiler for better .fasl-s
Instead of wrapping each toplevel expression in a lambda, we're just
compiling them as in a PROGN (discarding the return value).  This is good as
the compiler can now remove side-effect-free toplevel forms; also we don't
have pointless FN([...]),EXEC() anymore, thus improving the size of the
output (not to mention we're not anymore consing an useless environment
frame for each toplevel expression).