simplify comp-block (much faster too, in general)
Since, in general, most functions won't use return-from, it's wasteful to
compile in block first, then if zero usage is detected recompile outside
block (almost all functions are compiled twice!). The alternative here is to
compile outside block first - which will succeed for most cases - and if
return-from is encountered and there's no block with that name, throw and
retry inside block.