Closette WIP
The bulk of the diff is mostly stylistic: indentation (that's automatic and
Ymacs isn't terribly smart about it), no quotes for keywords or lambdas.

THE important change here is using a simple string key for effective method
caches, since our hash tables don't support EQUAL test.

It works! If you do (defclass foo () ()) it'll throw a stack overflow
exception (eventually), but that only happens when printing the result,
because it's a cyclic structure — my bad for not yet supporting
print-function for structures; I've removed that bit from Closette, but I
now understand why it was important.