Closette (defmethod): avoid COMPILE at load time
Previously, it stored code in method-body, and compiled it later in
compute-method-function, by calling COMPILE. Now it directly produces the
method function in the DEFMETHOD expansion.

It increases the binary size (e.g. printer.fasl, which defines 17 methods,
gained 2.3Kb); however, it decreases the total boot time by a factor of 2,
and we don't even have that many methods.